Created Project B file
Added 8 to 1 mux from old homework (still needs to be edited to be 16 wide) as well as the 2 to 1 and 8 to 1 mux modules to make the 3W to work. Also added decoder from old homework to decode info into HEX for the output. Finally, created the ProjectB top level module.